+++ src/BabelFish/DataModel/Clubs/ClubDetail.cs
@@ -1,4 +1,5 @@
 using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 using Scopos.BabelFish.Converters.Microsoft;
 using Scopos.BabelFish.DataModel.Common;
@@ -12,6 +13,22 @@ namespace Scopos.BabelFish.DataModel.Clubs {
         #region Private and Protected Fields
         private static Logger _logger = LogManager.GetCurrentClassLogger();
         private DateTime _memberSince = DateTime.Today;
+        private VisibilityOption _visibility = VisibilityOption.PUBLIC;
+
+        /// <summary>
+        /// Helper property to return the list of valid values for <see cref="Visibility"/>.
+        /// <list type="bullet">
+        /// <item>
+        /// <description>PROTECTED: May be seen by Club Members.</description>
+        /// </item>
+        /// <item>
+        /// <description>PUBLIC: May be seen by anyone.</description>
+        /// </item>
+        /// </list>
+        /// </summary>
+        [G_NS.JsonIgnore]
+        [G_STJ_SER.JsonIgnore]
+        public static readonly List<VisibilityOption> VisibilityOptions = new List<VisibilityOption>() { VisibilityOption.PROTECTED, VisibilityOption.PUBLIC };
         #endregion 
 
         #region Constructors, Factory Methods, and Initialization
@@ -29,6 +46,18 @@ namespace Scopos.BabelFish.DataModel.Clubs {
                 if (address == null) continue;
                 address.Club = this;
             }
+
+            // Set the backwards pointer for each Contact List
+            foreach (var contact in ContactList) {
+                if (contact == null) continue;
+                contact.Club = this;
+            }
+
+            // Set the backwards pointer for each License List
+            foreach (var license in LicenseList) {
+                if (license == null) continue;
+                license.Club = this;
+            }
         }
         #endregion
 
@@ -38,6 +67,7 @@ namespace Scopos.BabelFish.DataModel.Clubs {
         /// </summary>
         /// <example>1234</example>
         [DefaultValue( 0 )]
+        [Range( 0, 999999, ErrorMessage = "AccountNumber must be between 0 and 999999." )]
         public int AccountNumber { get; set; }
 
         /// <summary>
@@ -45,6 +75,8 @@ namespace Scopos.BabelFish.DataModel.Clubs {
         /// </summary>
         /// <example>Northeast High School</example>
         [DefaultValue( "" )]
+        [Required]
+        [StringLength( 128, ErrorMessage = "Name cannot be longer than 128 characters." )]
         public string Name { get; set; } = string.Empty;
 
         /// <summary>
@@ -56,6 +88,11 @@ namespace Scopos.BabelFish.DataModel.Clubs {
                 return $"OrionAcct{AccountNumber:D6}";
             }
         }
+
+        /// <summary>
+        /// Specifies if this is a INDIVIDUAL, HOME, or SITE license. This is used to determine what features are available to the club.
+        /// </summary>
+        public ClubLicenseType AccountType { get; set; } = ClubLicenseType.INDIVIDUAL;
 
         /// <summary>
         /// The list of people who are Administrators for this club.
@@ -106,12 +143,29 @@ namespace Scopos.BabelFish.DataModel.Clubs {
         }
 
         /// <summary>
-        /// The city and state (and maybe country) where the club is from.
+        /// Returns the hometown of the club, which is determined by the first address in the AddressList that has IsRange set to true.
+        /// If no such address exists, it will return the first physical address. If neither exists, it will return an empty string.
         /// </summary>
         /// <example>Axtell, NE</example>
         [DefaultValue( "" )]
-        [Obsolete( "Replaced with AddressList. June 2026." )]
-        public string Hometown { get; set; } = string.Empty;
+        public string Hometown {
+            get {
+                // Find the ClubAddress with IsRange set to true.
+                var rangeAddress = AddressList.FirstOrDefault( a => a.IsRange );
+                if (rangeAddress != null) {
+                    return StringFormatting.Hometown( rangeAddress.City, rangeAddress.State, rangeAddress.CountryCode );
+                }
+
+                // If no range address is found, find the first physical address.
+                var physicalAddress = AddressList.FirstOrDefault( a => a.IsPhysical );
+                if (physicalAddress != null) {
+                    return StringFormatting.Hometown( physicalAddress.City, physicalAddress.State, physicalAddress.CountryCode );
+                }
+
+                // While it would be unusual, a Club is not required to have any addresses, so if none are found, return an empty string.
+                return string.Empty;
+            }
+        }
 
         [Obsolete( "Replaced with AddressList. June 2026." )]
         public string Street1 { get; set; }
@@ -145,27 +199,40 @@ namespace Scopos.BabelFish.DataModel.Clubs {
         /// </summary>
         /// <example>northeast</example>
         [DefaultValue( "" )]
+        [Required]
+        [StringLength( 45, ErrorMessage = "URLPath cannot be longer than 45 characters." )]
         public string URLPath { get; set; } = string.Empty;
 
+
         /// <summary>
-        /// Gets or sets the visibility of this club's team page on rezults.scopos.net. It is the responsibility of a
-        /// Club's administrators or manager to set this value appropriately.
-        /// <para>If PRIVATE, the team page will not be visible to the public.</para>
-        /// <para>If PUBLIC, the team page will likely be visible to the public. In order to be visility, the Club must
-        /// have a valid Orion for Clubs license. Check <see cref="IsPublicUrlPageVisible"/> to learn if the Club
-        /// passes these tests.</para>
+        /// Gets or sets the visibility level that controls who can view this address.
+        /// The set value must be contained in <see cref="VisibilityOptions"/>. If not, the most restrictive option (the first in the list) will be used instead.
         /// </summary>
-        [G_NS.JsonProperty( DefaultValueHandling = G_NS.DefaultValueHandling.Include )]
-        public VisibilityOption Visibility { get; set; } = VisibilityOption.PRIVATE;
+        [G_STJ_SER.JsonPropertyOrder( 3 )]
+        [G_NS.JsonProperty( Order = 3, DefaultValueHandling = G_NS.DefaultValueHandling.Include )]
+        public VisibilityOption Visibility {
+            get => _visibility;
+            set {
+                if (!VisibilityOptions.Contains( value ))
+                    _visibility = VisibilityOptions[0];
+
+                _visibility = value;
+            }
+        }
+
         /// <summary>
         /// The x-api-key for use by this Club.
+        /// <para>As of June 2026 returns a fake placeholder value.</para>
         /// </summary>
-        public string ApiKey { get; set; } = string.Empty;
+        [Obsolete( "No longer used as of Orion version 2.25.7" )]
+        public string ApiKey { get { return "1234567890ABCDEFG"; } }
 
         /// <remarks>
         /// Should not be returned as part of the REST API response, in either public or authenticated calls.
+        /// <para>As of June 2026 returns a fake placeholder value.</para>
         /// </remarks>
-        public string ApiKeyId { get; set; } = string.Empty;
+        [Obsolete( "No longer used as of Orion version 2.25.7 / June 2026" )]
+        public string ApiKeyId { get { return "1234567890ABCDEFG"; } }
 
         /// <remarks>
         /// Should not be returned as part of the REST API response, in either public or authenticated calls.
@@ -180,6 +247,7 @@ namespace Scopos.BabelFish.DataModel.Clubs {
         /// <remarks>
         /// Should not be returned as part of the REST API response, in either public or authenticated calls.
         /// </remarks>
+        [Obsolete( "No longer used as of June 2026" )]
         public string AWSRegion { get; set; } = string.Empty;
 
         /// <summary>
@@ -188,7 +256,7 @@ namespace Scopos.BabelFish.DataModel.Clubs {
         public List<string> Notes { get; set; } = new List<string>();
 
         /// <summary>
-        /// The list of Orion Licenses this Club has. Most Clubs will have exactly one license. 
+        /// The list of Orion Licenses this Club has. Most Clubs will have exactly one license. Orion at Home accounts can have exactly one.
         /// </summary>
         public List<ClubLicense> LicenseList { get; set; } = new List<ClubLicense>();
 
@@ -202,12 +270,41 @@ namespace Scopos.BabelFish.DataModel.Clubs {
         #endregion
 
         #region Helper Properties
+
         /// <summary>
-        /// Helper property to return the list of valid VisibilityOption values. This is not returned as part of the REST API response, but is provided for ease of use in client applications.
+        /// Helper property to determine if this club may add another license. This is true if the club has no licenses or if the club has an Individual license. It is false if the club has a Site or Home license.
         /// </summary>
         [G_NS.JsonIgnore]
         [G_STJ_SER.JsonIgnore]
-        public static List<VisibilityOption> VisibilityOptions { get; private set; } = new List<VisibilityOption>() { VisibilityOption.PRIVATE, VisibilityOption.PUBLIC };
+        public bool MayAddLicense {
+            get {
+                // Site Licenses and Home Licenses can only have one license. Individual Licenses can have multiple licenses.
+                if (this.AccountType == ClubLicenseType.SITE || this.AccountType == ClubLicenseType.HOME) {
+                    return LicenseList.Count < 1;
+                }
+
+                return true;
+            }
+        }
+
+        /// <summary>
+        /// The list of roles that may be assigned to members of this Orion Account. The list of roles that is returned
+        /// depends on the AccountType. For SITE and INDIVIDUAL accounts, all roles except HOME_USER are available. For HOME accounts,
+        /// only the HOME_USER role is available. If the AccountType is not set to a valid value, an empty list is returned.
+        /// </summary>
+        /// <returns></returns>
+        public List<ClubAuthorizationRole> GetApplicableAuthorizationRoles() {
+            if (this.AccountType == ClubLicenseType.INDIVIDUAL || this.AccountType == ClubLicenseType.SITE) {
+                return new List<ClubAuthorizationRole>() { ClubAuthorizationRole.ADMIN, ClubAuthorizationRole.MANAGER, ClubAuthorizationRole.MEMBER, ClubAuthorizationRole.COACH, ClubAuthorizationRole.TECHNICAL_OFFICER, ClubAuthorizationRole.PAYER };
+            } else if (this.AccountType == ClubLicenseType.HOME) {
+                return new List<ClubAuthorizationRole>() { ClubAuthorizationRole.HOME_USER };
+            } else {
+                // Would only get here if the AccountType is the deprecated ClubLicenseType.TEMPORARY or if the AccountType is not set to a valid value. In either case, return an empty list.
+                _logger.Warn( $"GetApplicableAuthorizationRoles() called for club {this.OwnerId} with unknown AccountType {this.AccountType}. Returning HOME_USER role." );
+                return new List<ClubAuthorizationRole>();
+            }
+        }
+
         #endregion
 
         #region Methods
@@ -218,6 +315,16 @@ namespace Scopos.BabelFish.DataModel.Clubs {
         /// <returns></returns>
         public bool IsPublicUrlPageVisible() {
             return Visibility == VisibilityOption.PUBLIC && LicenseList.Any( l => (l.LicenseType == ClubLicenseType.INDIVIDUAL || l.LicenseType == ClubLicenseType.SITE) && l.ExpirationDate >= DateTime.Today );
+        }
+
+        /// <summary>
+        /// Returns true if Club Members, Admins, and Managers should be able to see the Club page page even if the club has not set its Visibility to PUBLIC.
+        /// This is true if the club has at least one valid Orion for Clubs license.
+        /// <para>Would be false if all of their licenses have expired, or this is a Orion at Home account.</para>
+        /// </summary>
+        /// <returns></returns>
+        public bool IsProtectedUrlPageVisible() {
+            return LicenseList.Any( l => (l.LicenseType == ClubLicenseType.INDIVIDUAL || l.LicenseType == ClubLicenseType.SITE) && l.ExpirationDate >= DateTime.Today );
         }
 
         /// <inheritdoc />