+++ src/BabelFish/DataModel/Clubs/ClubEnums.cs
@@ -8,6 +8,12 @@ namespace Scopos.BabelFish.DataModel.Clubs {
      * When adding new enums, it is usually best to update Helpers/SerializationOptions.cs
      */
 
+    /// <summary>
+    /// Defines the roles that can be assigned to members of a Club (aka Orion Account). The roles are used to
+    /// control access to various features and permissions within the system. This enum defines
+    /// roles for both Orion for Clubs and Orion at Home users. To get a list of roles that are specific to the
+    /// Club use <see cref="ClubDetail.GetApplicableAuthorizationRoles"/>.
+    /// </summary>
     public enum ClubAuthorizationRole {
         /*
          * These values should align with the values defined in the sql 'role' table.
@@ -39,7 +45,14 @@ namespace Scopos.BabelFish.DataModel.Clubs {
 
         [Description( "Club Technical Officer" )]
         [EnumMember( Value = "Club Technical Officer" )]
-        TECHNICAL_OFFICER
+        TECHNICAL_OFFICER,
+
+        /// <summary>
+        /// Specific role for the owner of an Orion at Home account. This role is not used for Orion for Clubs accounts.
+        /// </summary>
+        [Description( "Club Home" )]
+        [EnumMember( Value = "Club Home" )]
+        HOME_USER,
     }
 
     /// <summary>
@@ -141,7 +154,7 @@ namespace Scopos.BabelFish.DataModel.Clubs {
     /// <summary>
     /// Medea called these LicenseFeature
     /// </summary>
-    [Obsolete( "This feature is no longer in use as of 2026-04. Also not included on LicenseFiles" )]
+    [Obsolete( "This feature is no longer in use as of 2026-04. Maintained for backward compatibility with older Orion software." )]
     public enum ClubLicenseCapability {
 
         /// <summary>
@@ -170,9 +183,19 @@ namespace Scopos.BabelFish.DataModel.Clubs {
         /// </summary>
         [Description( "Privileged" )]
         [EnumMember( Value = "PRIVILEGED" )]
-        PRIVILEGED
+        PRIVILEGED,
+
+        ORION_V1,
+
+        ORION_HOME,
+
+        ORION_VS
     };
 
+    /// <summary>
+    /// Specifies the type of Orion Account (aka Club) license.
+    /// </summary>
+    /// <remarks>Value gets store to the orion_account sql table.</remarks>
     public enum ClubLicenseType {
 
         /// <summary>
@@ -199,6 +222,7 @@ namespace Scopos.BabelFish.DataModel.Clubs {
         /// <summary>
         /// Temporary Orion for Clubs license.
         /// </summary>
+        [Obsolete( "This license type is no longer in use as of 2026-04." )]
         [Description( "TEMPORARY" )]
         [EnumMember( Value = "TEMPORARY" )]
         TEMPORARY